Move rust code back to wasmer_vm module
