+++ wasm/src/script.c
@@ -1,10 +1,10 @@
 #include <ecs/components/Transform.h>
 
-extern void hello(Transform vec);
+extern void print_transform(Transform vec);
 
 Transform add(Transform a, Transform b) {
-    hello(a);
-    hello(b);
+    print_transform(a);
+    print_transform(b);
     GlmVec3 pos = {5, 6, 7};
     transform_set_position(&a, &pos);
     transform_get_position(&pos, &b);
