+++ wasm/src/test.c
@@ -1,9 +1,6 @@
 #include <ecs.h>
 
-extern void hello();
-
 int add(int a, int b) {
-    hello();
     return a + b;
 }
