+++ wasm/wasm.syms
@@ -1,4 +1,4 @@
-hello
+print_transform
 transform_identity
 transform_from_pos
 transform_get_position
